Restructured header component
