+++ src/components/Header/HeaderIcons/HeaderButtonWrapper/HeaderButtonWrapper.module.scss
@@ -1,4 +1,4 @@
-@import "../../scss/variables.scss";
+@import "../../../scss/variables.scss";
 
 .headerButton {
   cursor: pointer;
